QPID-2857 : Address issues found by running FindBugs against the Java codebase
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid@996393 13f79535-47bb-0310-9956-ffa450edef68

+++ java/systests/src/main/java/org/apache/qpid/test/client/failover/FailoverTest.java
@@ -306,7 +306,7 @@ public class FailoverTest extends FailoverBaseCase implements ConnectionListener
         //Start the connection
         connection.start();
 
-        long FAILOVER_DELAY = (RETRIES * DELAY);
+        long FAILOVER_DELAY = ((long)RETRIES * (long)DELAY);
 
         // Use Nano seconds as it is more accurate for comparision.
         long failTime = System.nanoTime() + FAILOVER_DELAY * 1000000;
